requirements.txt
supabase
python-dotenv
psycopg2-binary
asyncpg